+++ bin/prepare-release.sh
@@ -113,6 +113,32 @@ run $SED -i.bak -E \
     "${WORKSPACE_ROOT}/${FILE}"
 run rm "${WORKSPACE_ROOT}/${FILE}.bak"
 
+### Qt sample app ###
+# Qt changes are necessary because QML requires that you add
+# version number along with the import statements.
+
+# This gets the version string without the patch version.
+GLEAN_VERSION_FOR_QML=$(node -p -e "'${NEW_VERSION}'.split('.').reverse().slice(1).reverse().join('.')")
+
+FILE=samples/qt-qml-app/main.qml
+run $SED -i.bak -E \
+    -e "s/import org.mozilla.Glean [0-9a-z.-]+/import org.mozilla.Glean \"${GLEAN_VERSION_FOR_QML}\";/" \
+    -e "s/import generated [0-9a-z.-]+/import generated \"${GLEAN_VERSION_FOR_QML}\";/" \
+    "${WORKSPACE_ROOT}/${FILE}"
+run rm "${WORKSPACE_ROOT}/${FILE}.bak"
+
+FILE=samples/qt-qml-app/README.md
+run $SED -i.bak -E \
+    -e "s/--option platform=qt --option version=[0-9a-z.-]+/--option platform=qt --option version=\"${GLEAN_VERSION_FOR_QML}\";/" \
+    "${WORKSPACE_ROOT}/${FILE}"
+run rm "${WORKSPACE_ROOT}/${FILE}.bak"
+
+FILE=.circleci/config.yml
+run $SED -i.bak -E \
+    -e "s/--option platform=qt --option version=[0-9a-z.-]+/--option platform=qt --option version=\"${GLEAN_VERSION_FOR_QML}\";/" \
+    "${WORKSPACE_ROOT}/${FILE}"
+run rm "${WORKSPACE_ROOT}/${FILE}.bak"
+
 echo "Everything prepared for v${NEW_VERSION}"
 echo
 echo "Changed files:"